Fix data-correction migration to work where no data needs fixing.

+++ db/migrate/20120524151543_ensure_documents_can_only_belong_to_a_single_organisation.rb
@@ -9,9 +9,11 @@ class EnsureDocumentsCanOnlyBelongToASingleOrganisation < ActiveRecord::Migratio
         AND duplicates.id < document_organisations.id)
     }).collect {|x| x["id"]}
 
-    execute %{
-      DELETE FROM document_organisations WHERE id IN (#{ids.join(", ")})
-    }
+    unless ids.empty?
+      execute %{
+        DELETE FROM document_organisations WHERE id IN (#{ids.join(", ")})
+      }
+    end
 
     # Prevent future duplicates
     remove_index :document_organisations, :document_id
